feat: moved useStyles hook to own file
