+++ app/hooks.ts
@@ -1,105 +1,14 @@
-import { useEffect, useMemo, useRef } from "react"
+import { useMemo } from "react"
 
-import { useStorage } from "@plasmohq/storage/hook"
-
-import { getMeasureValue } from "~app/utils"
-
-import { loadFont } from "./api/fonts"
 import {
   GOAL_WIDGET,
   GW_IMAGE,
   GW_PROGRESS_BAR,
-  K_REM,
   LEFT_TEXT,
   RIGHT_TEXT
 } from "./constants"
-import useDebounce from "./hooks/useDebounce"
-import useDefaultStyles from "./hooks/useDefaultStyles"
-import useElements from "./hooks/useElements"
-import useFonts from "./hooks/useFonts"
-import type { SelectorProps, StylesData } from "./types"
-
-export const useStyles = () => {
-  const elements = useElements()
-  const defaultStyles = useDefaultStyles()
-
-  const isUpdatedStyles = useRef(false)
-
-  const { selectedFont } = useFonts()
-
-  useEffect(() => {
-    async function loadFonts() {
-      const fonts = Object.values(selectedFont)
-        .flatMap(({ value }) => value)
-        .filter(Boolean)
-
-      await Promise.all(fonts.map((font) => loadFont(font)))
-    }
-
-    if (selectedFont) {
-      loadFonts()
-    }
-  }, [selectedFont])
-
-  const [styles, setStyles] = useStorage("customStyles", (value?: StylesData) =>
-    value
-      ? Object.keys(defaultStyles).reduce(
-          (accum, selector) => ({
-            ...accum,
-            [selector]: { ...defaultStyles[selector], ...value[selector] }
-          }),
-          defaultStyles
-        )
-      : defaultStyles
-  )
-
-  const debouncedSetStyles = useDebounce(setStyles, 200)
-
-  const updElementStyles = (selector: string, prop: string, value: string) => {
-    let newValue = value
-    if (value.match("rem")) {
-      const { numberValue, measureValue } = getMeasureValue(value)
-
-      newValue = `${numberValue / K_REM}${measureValue}`
-    }
-
-    elements[selector].style[prop] = newValue
-  }
-
-  const updateStyles = (selector: string, prop: string, value: string) => {
-    debouncedSetStyles({
-      ...styles,
-      [selector]: { ...styles[selector], [prop]: value }
-    })
-
-    updElementStyles(selector, prop, value)
-  }
-
-  useEffect(() => {
-    if (isUpdatedStyles.current) {
-      return
-    }
-
-    for (const selector in styles) {
-      for (const prop in styles[selector]) {
-        updElementStyles(selector, prop, styles[selector][prop])
-      }
-    }
-
-    isUpdatedStyles.current = styles !== defaultStyles
-  }, [styles])
-
-  const resetStyles = () => {
-    setStyles(defaultStyles)
-    isUpdatedStyles.current = false
-  }
-
-  return {
-    styles,
-    resetStyles,
-    updateStyles
-  }
-}
+import useStyles from "./hooks/useStyles"
+import type { SelectorProps } from "./types"
 
 export const useData = () => {
   const { styles } = useStyles()
